refactor #239: move data calculation to helper
